Changes to the libuv to support forking.
This is pretty much the same thing as with the old node
version. I basically manually reapplied changes listed below.
They mostly deal with the fact that libuv uses epoll and some weird
machinery to route signals which needs to be reinitialized after
forking.

This time V8 didn't require any changes (but the profiler will not work
in the forked process).

Here are the changes that I applied:

commit 2d8d6f0bbe185f791d16935b7fcff7366e079327
Author: Konstantin Azarov <konstantin@stag-bastion000.ec2>
Date:   Sat Sep 28 00:32:03 2013 +0000

    Make pipes opened with uv_pipe_open nonblocking

commit fa237389e8499a0f55d7e442fee8ff9696d0982b
Author: Konstantin Azarov <konstantin.azarov@gmail.com>
Date:   Tue Jul 30 10:44:56 2013 +0000

    Reinit async after fork.

commit abaf96194eb9c31a35f6216429d468506e7b8348
Author: Konstantin Azarov <konstantin.azarov@gmail.com>
Date:   Mon Jul 29 13:06:45 2013 +0000

    More forking fixes:

      o Reinitialize epoll machinery in the child
      o Recreate pipes used for signal handling in the child

commit fda1d04c79d459bdaa8c108a9f8d71d6a9a6026f
Author: Konstantin Azarov <konstantin.azarov@gmail.com>
Date:   Tue Jul 23 08:15:16 2013 +0000

    Add functionality to suspend worker threads